+++ CHANGELOG.md
@@ -25,6 +25,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.15.1] - 2022-04-18
+
+### Fixed
+- Fixed tab completion in ``sat bash``.
+- Fixed a bug with the ``$PATH`` environment variable not including the ``sat``
+  executable in ``sat bash``.
+- Fixed error reporting in ``sat firmware`` to treat responses containing the
+  key "error" but an empty string for a value as non-errors.
+
 ## [3.15.0] - 2022-03-29
 
 ### Added